services/news_service/requirements.txt
django
djangorestframework
django-cors-headers  # Если нужен CORS